+++ internal/config/config.go
@@ -61,8 +61,10 @@ type Config struct {
 	notificationClientErr error
 	TimeZone              string
 
-	Ma *jsonformat.Marshaller
-	Um *jsonformat.Unmarshaller
+	STU3MA *jsonformat.Marshaller
+	STU3UM *jsonformat.Unmarshaller
+	R4MA   *jsonformat.Marshaller
+	R4UM   *jsonformat.Unmarshaller
 }
 
 func (c *Config) IAMClient() (*iam.Client, error) {
@@ -127,7 +129,7 @@ func (c *Config) SetupIAMClient() {
 	c.iamClient = nil
 	client, err := iam.NewClient(standardClient, &c.Config)
 	if err != nil {
-		c.iamClientErr = err
+		c.iamClientErr = fmt.Errorf("possible invalid region/environment: %w", err)
 		return
 	}
 	if c.ServiceID != "" && c.ServicePrivateKey != "" {