+++ internal/services/cdr/org/resource_cdr_org_v0.go
@@ -1,0 +1,48 @@
+package org
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// Upgrades a CDR Org from v0 to v1
+func patchCDROrgV0(ctx context.Context, rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
+	if rawState == nil {
+		rawState = map[string]interface{}{}
+	}
+	rawState["version"] = "stu3"
+	return rawState, nil
+}
+
+func resourceCDROrgV0() *schema.Resource {
+	return &schema.Resource{
+		// This is only used for state migration, so the CRUD
+		// callbacks are no longer relevant
+		Schema: map[string]*schema.Schema{
+			"fhir_store": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"org_id": {
+				Type:     schema.TypeString,
+				Required: true,
+				ForceNew: true,
+			},
+			"name": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"part_of": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"purge_delete": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Default:  false,
+			},
+		},
+	}
+}