+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.25.0
+
+- CDR: Support STU3 and R4 FHIR resources. Defaults to STU3, no change to existing resources
+- Provider: Add validation to region and environment fields (#54)
+- IAM Service: Mark expires_on as computed (#94)
+
 ## v0.24.1
 
 - Made `region` optional and default to `us-east`